Change before/after to be raw JSON
This optimizes for the case where events are seralized to JSON and then
sent elsewhere rather than evaluated within the engine. I expect this
be the common case.

Defering evaluation of the JSON results in an order of magnitude less
allocations in all benchmarks and a range of 10% - 25% reduction in
runtime across all benchmarks.